Add tests for socials
