+++ packages/server/customer-os-api/graph/resolver/organization.resolvers_it_test.go
@@ -1167,16 +1167,14 @@ func TestMutationResolver_OrganizationAddNewLocation(t *testing.T) {
 
 	organizationId := neo4jt.CreateOrganization(ctx, driver, tenantName, "org name")
 
-	rawResponse, err := c.RawPost(getQuery("organization/add_new_location_to_organization"),
-		client.Var("organizationId", organizationId),
-	)
-	assertRawResponseSuccess(t, rawResponse, err)
+	rawResponse := callGraphQL(t, "organization/add_new_location_to_organization",
+		map[string]interface{}{"organizationId": organizationId})
 
 	var organizationStruct struct {
 		Organization_AddNewLocation model.Location
 	}
 
-	err = decode.Decode(rawResponse.Data.(map[string]any), &organizationStruct)
+	err := decode.Decode(rawResponse.Data.(map[string]any), &organizationStruct)
 	require.Nil(t, err)
 	require.NotNil(t, organizationStruct)
 	location := organizationStruct.Organization_AddNewLocation
@@ -1191,4 +1189,54 @@ func TestMutationResolver_OrganizationAddNewLocation(t *testing.T) {
 	require.Equal(t, 1, neo4jt.GetCountOfNodes(ctx, driver, "Location"))
 	require.Equal(t, 1, neo4jt.GetCountOfRelationships(ctx, driver, "ASSOCIATED_WITH"))
 	assertNeo4jLabels(ctx, t, driver, []string{"Tenant", "Location", "Location_" + tenantName, "Organization", "Organization_" + tenantName})
+}
+
+func TestQueryResolver_Organization_WithSocials(t *testing.T) {
+	ctx := context.TODO()
+	defer tearDownTestCase(ctx)(t)
+	neo4jt.CreateTenant(ctx, driver, tenantName)
+	orgId := neo4jt.CreateOrganization(ctx, driver, tenantName, "org name")
+
+	socialId1 := neo4jt.CreateSocial(ctx, driver, tenantName, entity.SocialEntity{
+		PlatformName: "p1",
+		Url:          "url1",
+	})
+	socialId2 := neo4jt.CreateSocial(ctx, driver, tenantName, entity.SocialEntity{
+		PlatformName: "p2",
+		Url:          "url2",
+	})
+	neo4jt.LinkSocialWithEntity(ctx, driver, orgId, socialId1)
+	neo4jt.LinkSocialWithEntity(ctx, driver, orgId, socialId2)
+
+	require.Equal(t, 1, neo4jt.GetCountOfNodes(ctx, driver, "Organization"))
+	require.Equal(t, 2, neo4jt.GetCountOfNodes(ctx, driver, "Social"))
+	require.Equal(t, 2, neo4jt.GetCountOfRelationships(ctx, driver, "HAS"))
+
+	rawResponse := callGraphQL(t, "organization/get_organization_with_socials",
+		map[string]interface{}{"organizationId": orgId})
+
+	var orgStruct struct {
+		Organization model.Organization
+	}
+
+	err := decode.Decode(rawResponse.Data.(map[string]any), &orgStruct)
+	require.Nil(t, err)
+
+	organization := orgStruct.Organization
+	require.NotNil(t, organization)
+	require.Equal(t, 2, len(organization.Socials))
+
+	require.Equal(t, socialId1, organization.Socials[0].ID)
+	require.Equal(t, "p1", *organization.Socials[0].PlatformName)
+	require.Equal(t, "url1", organization.Socials[0].URL)
+	require.NotNil(t, organization.Socials[0].CreatedAt)
+	require.NotNil(t, organization.Socials[0].UpdatedAt)
+	require.Equal(t, "test", organization.Socials[0].AppSource)
+
+	require.Equal(t, socialId2, organization.Socials[1].ID)
+	require.Equal(t, "p2", *organization.Socials[1].PlatformName)
+	require.Equal(t, "url2", organization.Socials[1].URL)
+	require.NotNil(t, organization.Socials[1].CreatedAt)
+	require.NotNil(t, organization.Socials[1].UpdatedAt)
+	require.Equal(t, "test", organization.Socials[1].AppSource)
 }
