+++ packages/server/customer-os-api/graph/resolver/test_queries/contact/add_social_to_contact.txt
@@ -1,0 +1,17 @@
+mutation AddSocialToContact($contactId: ID!) {
+  contact_AddSocial(
+        contactId: $contactId,
+        input: {
+            url: "social url"
+            platformName: "social platform"
+        }) {
+    id
+    createdAt
+    updatedAt
+    appSource
+    source
+    sourceOfTruth
+    url
+    platformName
+  }
+}
