+++ packages/server/customer-os-api/graph/resolver/test_queries/contact/get_contact_with_socials.txt
@@ -1,0 +1,13 @@
+query GetContact($contactId: ID!){
+  contact(id: $contactId) {
+    id
+    socials {
+            id
+            url
+            platformName
+            createdAt
+            updatedAt
+            appSource
+          }
+  }
+}
