+++ packages/server/customer-os-api/graph/resolver/test_queries/organization/get_organization_with_socials.txt
@@ -1,0 +1,13 @@
+query GetOrganization($organizationId: ID!){
+  organization(id: $organizationId) {
+    id
+    socials {
+            id
+            url
+            platformName
+            createdAt
+            updatedAt
+            appSource
+          }
+  }
+}
