+++ packages/server/customer-os-api/graph/resolver/test_queries/social/update_social.txt
@@ -1,0 +1,15 @@
+mutation UpdateSocial($socialId: ID!) {
+  social_Update(
+    input: {
+      id: $socialId
+      platformName: "new name"
+      url: "new url"
+    }
+  ) {
+    id
+    updatedAt
+    sourceOfTruth
+    platformName
+    url
+  }
+}
