+++ packages/server/customer-os-api/graph/schemas/contact.graphqls
@@ -41,7 +41,7 @@ extend type Mutation {
     contact_RemoveOrganizationById(input: ContactOrganizationInput!): Contact!
 
     contact_AddNewLocation(contactId: ID!): Location!
-    contact_AddSocial(contactId: ID!, input: SocialInput): Social!
+    contact_AddSocial(contactId: ID!, input: SocialInput!): Social!
 }
 
 """
